requirements.txt
requests==2.32.3
python-dotenv==1.0.1
pandas==2.2.2
pyarrow==16.1.0
pyspark==3.5.1
pendulum==3.0.0